\name{cormorant}
\alias{cormorant}
\title{Cormorant algorithm}
\usage{
  cormorant(tumorFiles, normalFiles, regions, nucleotides,
    mc.cores = 1, q = 25, ...)
}
\arguments{
  \item{tumorFiles}{}

  \item{normalFiles}{}

  \item{regions}{}

  \item{nucleotides}{}

  \item{mc.cores}{}

  \item{q}{}

  \item{...}{}
}
\value{
  A data.frame with p-values
}
\description{
  Cormorant algorithm
}
\author{
  mg14
}

