% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSNV-package.R
\docType{data}
\name{counts}
\alias{counts}
\title{Example count table}
\description{
A table with counts of the HIVmix data set. Used for minimal unit testing.
}
\examples{
data("counts", package="deepSNV")
countsFromBam <- bam2R(file = system.file("extdata", "test.bam", package="deepSNV"), chr="B.FR.83.HXB2_LAI_IIIB_BRU_K034", start = 3120, stop=3140, q = 10)
all(counts == countsFromBam)
}
