\name{dbbinom}
\alias{dbbinom}
\title{Beta-binomial distribution}
\usage{
  dbbinom(x, n, mu, rho, log = FALSE)
}
\arguments{
  \item{x}{Counts, integer}

  \item{n}{Size, integer}

  \item{mu}{Probability, nonnegative numeric}

  \item{rho}{Dispersion, numeric (0,1)}

  \item{log}{Logical. Should logarithmic probabilities be
  returned?}
}
\value{
  Probability
}
\description{
  Beta-binomial distribution
}
\author{
  mg14
}

