\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-binomial probability distribution}
\usage{
  dbetabinom(x, size, prob, dispersion, log = FALSE)
}
\arguments{
  \item{x}{Counts}

  \item{size}{Size}

  \item{prob}{Probability}

  \item{dispersion}{Dispersion rho/(1-rho)}

  \item{log}{Should the natural logarithm be returned?
  Default=FALSE}
}
\value{
  Probability
}
\description{
  Beta-binomial probability distribution
}
\author{
  mg14
}

