% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shearwater-functions.R
\name{loadAllData}
\alias{loadAllData}
\title{Function to load all data from a list of bam files}
\usage{
loadAllData(files, regions, ..., mc.cores = 1)
}
\arguments{
\item{files}{A character vector with the paths to all bam files}

\item{regions}{Either a GRanges or data.frame with the coordinates of interest}

\item{...}{Arguments passed to bam2R}

\item{mc.cores}{Number of cores used for loading, default = 1}
}
\value{
counts
}
\description{
This function uses the parallel package and the bam2R interface to load all nucleotide counts from a list of bam files and a set of regions into a large array.
}
\note{
Experimental code, subject to changes
}
\author{
mg14
}

