\name{pbbinom}
\alias{pbbinom}
\title{Beta-binomial cumulative density}
\usage{
  pbbinom(q, n, mu, rho, lower.tail = FALSE, log.p = FALSE)
}
\arguments{
  \item{q}{Counts, integer}

  \item{n}{Size, integer}

  \item{mu}{Probability, nonnegative numeric}

  \item{rho}{Dispersion, numeric (0,1)}

  \item{log.p}{Logical. Should logarithmic probabilities be
  returned?}
}
\value{
  Cumulative probability
}
\description{
  Beta-binomial cumulative density
}
\author{
  mg14
}

