% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepSNV-package.R
\docType{data}
\name{phiX}
\alias{phiX}
\title{Example phiX data}
\description{
Data from two phiX experiments sequenced on a GAIIx.
}
\examples{
data(phiX, package="deepSNV")
plot(phiX)
phiN <- normalize(phiX, round=TRUE)
plot(phiN)
}

