% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shearwater-ML.R
\name{pvals_betabinLRT}
\alias{pvals_betabinLRT}
\title{Shearwater ML}
\usage{
pvals_betabinLRT(counts, rho = NULL, truncate = 0.01, rho.min = 1e-04,
  rho.max = 0.8)
}
\arguments{
\item{counts}{The array of counts typically generated by loadAllData.}

\item{rho}{A number}

\item{truncate}{}

\item{rho.min}{}
}
\value{
rho.max
}
\description{
This is shearwater generating P-values
}
\examples{
# code to be added
}
\author{
Inigo Martincorena
}

