% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betabinom.R
\name{dbetabinom}
\alias{dbetabinom}
\title{Beta-binomial probability distribution}
\usage{
dbetabinom(x, n, mu, rho, log = FALSE)
}
\arguments{
\item{x}{Counts}

\item{n}{Size}

\item{mu}{Probability}

\item{rho}{Dispersion. rho in (0,1)}

\item{log}{Return logarithmic values}
}
\value{
d
}
\description{
Beta-binomial probability distribution
}
\author{
mg14
}

