% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deepSNV-generics.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[,deepSNV,ANY,ANY-method}
\title{Subsetting for deepSNV objects.}
\usage{
\S4method{[}{deepSNV,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{A \code{\link{deepSNV-class}} object.}

\item{i}{Row indeces.}

\item{j}{Column (nucleotide) indeces.}
}
\value{
A \code{\link{deepSNV-class}} object.
}
\description{
Subsetting for deepSNV objects.
}
\examples{
data(HIVmix)
HIVmix[1:10,]
}
\author{
Moritz Gerstung
}

