% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{p.val}
\alias{p.val}
\alias{p.val,deepSNV-method}
\title{Get p-values}
\usage{
p.val(deepSNV, ...)

\S4method{p.val}{deepSNV}(deepSNV)
}
\arguments{
\item{deepSNV}{a \code{\link{deepSNV-class}} object}

\item{...}{Additional param passed to specific methods}
}
\value{
A matrix with the p-values.
}
\description{
Convenience function to get the p-values from a deepSNV object.
}
\examples{
data(HIVmix)
p.val(HIVmix)[1:10,]
}

