% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/betabinom.R
\name{pbetabinom}
\alias{pbetabinom}
\title{Cumulative beta-binomial probability distribution}
\usage{
pbetabinom(x, n, mu, rho, log = FALSE)
}
\arguments{
\item{x}{Counts}

\item{n}{Sample size}

\item{mu}{Probability}

\item{rho}{Dispersion. rho in (0,1)}

\item{log}{Return logarithmic values}
}
\value{
Probability
}
\description{
Cumulative beta-binomial probability distribution
}
\author{
mg14
}

